+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Upgraded to Laravel 8.0 from Laravel 7.0
 - Upgraded to Laravel 7.0 from Laravel 6.0
 
+### Fixed
+- Redirect to module includes query string
+
 ### Added
 - Wrapper around body of frontend.
 - Wrapper around content